Moving the tabular view into its own widget class. Unfinished.
